% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleansing.R
\name{Clean_Umlaute2}
\alias{Clean_Umlaute2}
\title{cleansing}
\usage{
Clean_Umlaute2(x)
}
\arguments{
\item{x}{string}
}
\description{
Data cleansing
}
\details{
Clean_Umlaute2(): Funktion entfernt stoerende Umlaute,
}
