% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countDigits.R
\name{countDigits}
\alias{countDigits}
\title{countDigits}
\usage{
countDigits(x)
}
\arguments{
\item{x}{zu pruefendes Objekt}
}
\value{
Anzahl an Nachkommastellen
}
\description{
Interne Function wird in Meanci2() verwendet
}
\examples{
countDigits(1.2345)
}
