% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateProjekt.R
\name{CreateProjekt}
\alias{CreateProjekt}
\title{Neues Projekt erstellen}
\usage{
CreateProjekt(project = "000 Dummy", datum = date(),
  comment = "Test Dummy", path = "C:/Users/wpete/Dropbox/1_Projekte")
}
\arguments{
\item{project}{Name des projektes}

\item{datum}{Datum}

\item{comment}{Beschreibung}

\item{path}{Pfad default Dropbox/1_Projekte}
}
\description{
Erstellt ein neues Projekt mir Ordnerstrucktur
 und R-Auswertungs-Files.
}
\examples{
## Not run:


#CreateProjekt( project= "001 Hans Meier",
#              datum= date(),
#              comment= "Auswertung mit Regressionsanalyse",
#              path="C:/Users/wpete/Dropbox/1_Projekte" )
## End(**Not run**)
}
