% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateProjekt.R
\name{folder}
\alias{folder}
\title{Create Folder}
\usage{
folder(..., folder.name = NULL)
}
\arguments{
\item{...}{Namen der Folder}

\item{folder.name}{weis nicht  = NULL}
}
\value{
\code{folder} creates a folder/directory.
}
\description{
\code{folder} - Create a folder/directory.
}
