% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_.R, R/zzz.R
\name{is_irgendwas}
\alias{is_irgendwas}
\alias{is_formula2}
\alias{is_empty2}
\alias{is_all_dichotom}
\alias{is_all_logical}
\alias{is_all_0_1}
\alias{isFALSE}
\alias{is_false2}
\alias{all_identical2}
\alias{is_all_identical2}
\alias{is_vars_in_data}
\alias{is_attached}
\title{is_ Familie oder as_Familie}
\usage{
is_formula2(x)

is_empty2(x)

is_all_dichotom(x)

is_all_logical(x)

is_all_0_1(x)

isFALSE(x)

is_false2(x)

all_identical2(x)

is_all_identical2(x)

is_vars_in_data(x, data = NULL)

is_attached(x)
}
\arguments{
\item{x}{zu pruefendes Objekt}

\item{data}{Daten wenn Formeln gepruft werden}
}
\value{
Die \code{is_all_} gibt generel einen Wert zurueck die \code{is_} einen Vector in gleicher Laenge wie der Input-Vector.
}
\description{
Prueft ob objekt bestimmte Eigenschaften aufweist.
Fuer Dataframe gibt es \code{is_all_identical2()}

is_formula2 Prueft ob es eine Formel ist

is_empty2 wird in prepare_data genutzt als test ob ein Elemen  leer ist

is_all_logical is_all_0_1 prufen beide Logical aber is_all_dichotom  kann auch ja/nein

isFALSE analog wie if(x){...}

is_false2 arbeitet mit isFALSE geht aber auch fuer Matris oder Data.frames

is_all_identical2 oder all_identical2 wird in Recast2 verwendet

is_vars_in_data Prueft ob ded data.frame auch die Fariablen enthaelt.
}
\examples{
# isFALSE(TRUE)
# x<-c(F, T, F, F)
# is_false2(x)
# is_all_logical(x)
# is_all_0_1(x)
# is_all_identical2(data.frame(y=1:3, x=factor(c("a", "b", "c"))))
# as_numeric(c(2, 2, 3, "Q))
 #


is_formula2(a~b+v)
 is_empty2(c("freq", "mean"))
 is_empty2("freq")
is_attached("stpvers")
}
