% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_.R
\name{as_irgenwas2}
\alias{as_irgenwas2}
\alias{as_numeric}
\title{as_Familie}
\usage{
as_numeric(x)
}
\arguments{
\item{x}{zu pruefendes Objekt}
}
\value{
Output ist ein Vector bzq dataframe mit gleicgher Laenge wie der Input.
}
\description{
Prueft ob objekt bestimmte Eigenschaften aufweist.
Fuer Dataframe gibt es \code{is_all_identical2()}

as_numeric fuer factor und string-objekte
}
\examples{
data<- data.frame(a=1:10,
  b=as.character(1:10),
c=gl(2, 5, labels = c("Control", "Treat")),
d=gl(2, 5, labels = c("1", "3"))

)
as_numeric(data$a)+1
as_numeric(data$b)+1
as_numeric(data$c)+1
as_numeric(data$d)+1
as_numeric(data)

x<- c("0", "-3",  " 5", " 5.6", "   3,5  ", "super")
character_to_numeric(factor(x))

x<- c("0", "-3",  " 5", " 5.6", "   3,5  ")
character_to_numeric(x)
}
