% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_.R
\name{is_irgendwas}
\alias{is_irgendwas}
\alias{is_formula2}
\alias{is_empty2}
\alias{is_false2}
\alias{is_all_identical2}
\alias{is_all_dichotom}
\alias{is_all_logical}
\alias{is_all_0_1}
\alias{isFALSE}
\alias{all_identical2}
\alias{is_vars_in_data}
\title{is_ Familie oder as_Familie}
\usage{
is_formula2(x)

is_empty2(x)

is_false2(x)

is_all_identical2(x)

is_all_dichotom(x)

is_all_logical(x)

is_all_0_1(x)

isFALSE(x)

all_identical2(data)

is_vars_in_data(x, data = NULL)
}
\arguments{
\item{x}{zu pruefendes Objekt}

\item{data}{Daten wenn Formeln geprüft werden}
}
\value{
Die \code{is_all_} gibt generel einen Wert zurueck die \code{is_} einen Vector in gleicher Laenge wie der Input-Vector.
}
\description{
Prueft ob objekt bestimmte Eigenschaften aufweist.
Fuer Dataframe gibt es \code{is_all_identical2()}

is empty wird in prepare_data genutzt als test ob ein Elemen  leer ist
}
\examples{
# isFALSE(TRUE)
# x<-c(F, T, F, F)
# is_false2(x)
# is_all_logical(x)
# is_all_0_1(x)
# is_all_identical2(data.frame(y=1:3, x=factor(c("a", "b", "c"))))
# as_numeric(c(2, 2, 3, "Q))
 #


is_formula2(a~b+v)
 is_empty2(c("freq", "mean"))
 is_empty2("freq")
}
